Moved table into its own component
